LINE INPUT control code fix #1
In all the dialects I've worked with I've never seen LINE INPUT alter
anything but EOL characters on input. The only thing it did with EOL
was stop and leave them out of the resulting string.

This is fixes that issue but it needs further scrutiny to see if it
breaks anything else and / or if a similar patch is needed somewhere
else.

I can now count the number of times a line is started with a tab!! 8-)